Check if SHA1 exists before calling `git diff`
In the case of a force push, the commit might not exist anymore, and the
`git diff` in `has_changes` will fail, but the failure would be silently
ignored (since the error message would be passed to `wc`).

`wc -l` would return 1, so effectively the same image would be reused
over and over.