Rename bazel to bazel-rules
The `bazel` executable installed by the official packaging is a shell script.
One of the things it does is look for `$PWD/tools/bazel` and try to run that,
presumably to allow for easy customization.

Since we vendor build-tools in as `tools/` almost everywhere,
we cannot have a directory named `bazel`
as the `bazel` CLI will try to execute `tools/bazel` and fail.