scheduler: pass CircleCI API token in for private projects
N.B.: This API token is currently attached to marc@weave.works and should be migrated to a functional account to avoid bad surprises in the future.
Also, the account used needs to follow the project for it to be accessible via the API. Otherwise, you may get some 404s.